Refactor and improve abscissa::testing
- Factors all testing types into their own submodules.
- Unifies `CmdRunner` and `CargoRunner` into the former.
- Import testing types as `abscissa::testing::prelude::*`.
- Adds an `exclusive()` method which serializes execution of commands
  under the same runner using a `Mutex`. Enabled by default when using
  `CmdRunner::default()`.
- Adds a `quiet()` method to `CmdRunner` to suppress logging information
  about executed commands.
