+++ abscissa_generator/template/tests/acceptance.rs.hbs
@@ -4,18 +4,31 @@
 //! For more information, see:
 //! <https://docs.rs/abscissa/latest/abscissa/testing/index.html>
 
-use abscissa::testing::CmdRunner;
+use abscissa::testing::prelude::*;
+use lazy_static::lazy_static;
+
+lazy_static! {
+    /// Executes your application binary via `cargo run`.
+    ///
+    /// Storing this value in a `lazy_static!` ensures that all instances of
+    /// the runner acquire a mutex when executing commands and inspecting
+    /// exit statuses, serializing what would otherwise be multithreaded
+    /// invocations as `cargo test` executes tests in parallel by default.
+    pub static ref RUNNER: CmdRunner = CmdRunner::default();
+}
 
 #[test]
 fn start_no_args() {
-    let mut cmd = CmdRunner::default().arg("start").capture_stdout().run();
+    let mut runner = RUNNER.clone();
+    let mut cmd = runner.arg("start").capture_stdout().run();
     cmd.stdout().expect_line("Hello, world!");
     cmd.wait().unwrap().expect_success();
 }
 
 #[test]
 fn start_with_args() {
-    let mut cmd = CmdRunner::default()
+    let mut runner = RUNNER.clone();
+    let mut cmd = runner
         .args(&["start", "acceptance", "test"])
         .capture_stdout()
         .run();
