+++ src/testing/prelude.rs
@@ -1,0 +1,6 @@
+//! Import prelude for Abscissa tests
+
+pub use super::{
+    process::{ExitStatus, OutputStream, Process, Stderr, Stdout},
+    runner::CmdRunner,
+};
